apps/api/requirements.txt
uvicorn[standard]==0.34.0
langchain==0.3.13
langchain-openai==0.2.14
langchain-google-genai==2.0.8
langgraph==0.2.60
fastapi==0.115.5
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic-settings==2.7.1
pytest==8.3.4
pytest-asyncio==0.24.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
python-multipart==0.0.20
redis
